transport_ocr_system/services/web-app/requirements.txt
# Django
Django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5

# PostgreSQL
psycopg2-binary==2.9.9

# MinIO
minio==7.2.0

# Утилиты
python-dotenv==1.0.0
Pillow==10.2.0
requests==2.31.0

# Celery
celery==5.3.6
redis==5.0.1

# Логирование
python-json-logger==2.0.7

# Мониторинг
prometheus-client==0.19.0
